requirements.txt
python-telegram-bot[job-queue]>=20.0
google-api-python-client>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.2.0
gspread>=6.0
groq>=0.4.0
python-dotenv>=1.0
yt-dlp>=2024.0
Flask>=3.0.0
requests>=2.31.0
